rename metrics module to usage_metrics
